Bug #2111497 sprd_bl u32 change to u64
[root cause  ]: u32 type compile failure
[changes     ]: u32 change to u64
[side effects]: none
[self test   ]: yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:none
[confirm dependent commit]:yes
[board]:all
[test case]:display case
[reviewers]:xiaoming.ren

commit_template_version:v1
[change_type ] debug_log
[tag_product ] common

+++ dispc/sprd_bl.c
@@ -18,7 +18,7 @@
 void sprd_backlight_normalize_map(struct backlight_device *bd, u16 *level)
 {
 	struct sprd_backlight *bl = bl_get_data(bd);
-	u32 mul;
+	u64 mul;
 
 	if (!bl->num) {
 		if(!bd->props.brightness)
